backend/pyproject.toml
# TaskMaster Pro - Backend Python Project Configuration
# Generated with GitHub Copilot assistance
# Date: 2026-03-12

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taskmaster-pro-backend"
version = "1.0.0"
description = "TaskMaster Pro backend API - FastAPI service for task management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "TaskMaster Pro Development Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "mysql-connector-python>=8.3.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.4",
    "httpx>=0.26.0",
    "pylint>=3.0.3",
    "black>=24.1.1",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
    "-v",
]
asyncio_mode = "auto"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]
max-line-length = 100
